update ids-splitter attribs/minor cleanup, add boilerplate for ids-splitter-pane
