+++ app/ids-splitter/example.html
@@ -1,1 +1,13 @@
-<ids-splitter></ids-splitter>
+<ids-layout-grid>
+  <ids-text font-size="12" type="h1">Splitter</ids-text>
+</ids-layout-grid>
+<ids-layout-grid>
+  <ids-splitter axis="x">
+    <ids-splitter-pane min-size="30%" max-size="50%">
+      some content in a child div
+    </ids-splitter-pane>
+    <ids-splitter-pane>
+      some content in a child div
+    </ids-splitter-pane>
+  </ids-splitter>
+</ids-layout-grid>
