+++ app/ids-splitter/index.js
@@ -1,1 +1,1 @@
-import IdsSplitter from '../../src/ids-splitter';
+import IdsSplitter, { IdsSplitterPane } from '../../src/ids-splitter';
