--- src/ids-splitter/ids-draggable.scss
@@ -1,7 +1,0 @@
-.ids-draggable {
-  display: flex;
-  width: 32px;
-  height: 100%;
-  background-color: #f00;
-  cursor: grab;
-}
