+++ src/ids-splitter/ids-splitter-pane.scss
@@ -1,0 +1,6 @@
+.ids-splitter-pane {
+  display: flex;
+  flex-direction: var(--axis-direction);
+  width: 100%;
+  height: 100%;
+}
