+++ src/ids-splitter/ids-splitter.scss
@@ -1,6 +1,5 @@
 .ids-splitter {
   display: flex;
-  width: 32px;
+  width: 100%;
   height: 100%;
-  background-color: #f00;
 }
