+++ src/ids-splitter/index.js
@@ -1,1 +1,2 @@
 export { default } from './ids-splitter';
+export { default as IdsSplitterPane } from './ids-splitter-pane';
